+++ pkg/controller/cloud/BUILD
@@ -11,7 +11,6 @@ go_library(
     srcs = ["node_controller.go"],
     deps = [
         "//pkg/api/v1/node:go_default_library",
-        "//pkg/client/retry:go_default_library",
         "//pkg/cloudprovider:go_default_library",
         "//pkg/kubelet/apis:go_default_library",
         "//pkg/util/node:go_default_library",
@@ -28,6 +27,7 @@ go_library(
         "//vendor/k8s.io/client-go/kubernetes/typed/core/v1:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
         "//vendor/k8s.io/client-go/tools/record:go_default_library",
+        "//vendor/k8s.io/client-go/util/retry:go_default_library",
     ],
 )
