+++ pkg/controller/history/BUILD
@@ -32,7 +32,6 @@ go_library(
     name = "go_default_library",
     srcs = ["controller_history.go"],
     deps = [
-        "//pkg/client/retry:go_default_library",
         "//pkg/util/hash:go_default_library",
         "//vendor/k8s.io/api/apps/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
@@ -47,6 +46,7 @@ go_library(
         "//vendor/k8s.io/client-go/kubernetes/scheme:go_default_library",
         "//vendor/k8s.io/client-go/listers/apps/v1beta1:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
+        "//vendor/k8s.io/client-go/util/retry:go_default_library",
     ],
 )
